switch Vector2 to float

+++ source/common/interface/vector2.hpp
@@ -5,24 +5,24 @@
 
 class Vector2 {
 public:
-  double x;
-  double y;
+  float x;
+  float y;
 
-  Vector2(double x = 0.0, double y = 0.0) : x(x), y(y){};
+  Vector2(float x = 0.0, float y = 0.0) : x(x), y(y){};
 
   // Scalar Operations
-  Vector2 operator+(const double scalar) const;
-  Vector2 operator*(const double scalar) const;
+  Vector2 operator+(const float scalar) const;
+  Vector2 operator*(const float scalar) const;
 
   // Vector Operations
   Vector2 operator+(const Vector2 &other) const;
   Vector2 operator*(const Vector2 &other) const;
-  double dot(const Vector2 &other);
+  float dot(const Vector2 &other);
 
   Vector2 &operator+=(const Vector2 &other);
-  Vector2 &operator*=(double scalar);
+  Vector2 &operator*=(float scalar);
 
-  double magnitude() const;
+  float magnitude() const;
   Vector2 normal() const;
 };
